+++ src/pdftl/cli/complete.py
@@ -12,9 +12,11 @@ import sys
 PICKLER = "cloudpickle"
 
 # if grammar.py output changes: update this!
-GRAMMAR_VERSION = "1"
+GRAMMAR_VERSION = "2"
 
 HARDCODED_KEYWORDS = [
+    "JOB",
+    "DONE",
     "add_text",
     "attach_files",
     "background",
@@ -213,9 +215,13 @@ def is_package_newer_than_cache(cache_path):
         # Add the folders where your registry/grammar logic actually lives.
         # e.g., if you add a new operation file in 'operations/', completion needs to know.
         if os.name == "nt":
-            custom_op_base = os.environ.get("APPDATA") or os.path.expanduser("~\\AppData\\Roaming")
+            custom_op_base = os.environ.get("APPDATA") or os.path.expanduser(
+                "~\\AppData\\Roaming"
+            )
         else:
-            custom_op_base = os.environ.get("XDG_CONFIG_HOME") or os.path.expanduser("~/.config")
+            custom_op_base = os.environ.get("XDG_CONFIG_HOME") or os.path.expanduser(
+                "~/.config"
+            )
 
         custom_op_dir = os.path.join(custom_op_base, "pdftl", "operations")
 
@@ -275,6 +281,8 @@ def resolve_candidates(allowed_tokens, parser):
     # Map for explicit terminals in GrammarBuilder
     # This must match the names in GrammarBuilder
     literal_map = {
+        "KW_SUB": "JOB",
+        "KW_END": "DONE",
         "HELP_KW": "help",
         "HELP_FLAG": "--help",
         "COMP_FLAG": "--completion",