+++ src/pdftl/cli/constants.py
@@ -10,3 +10,5 @@ DEBUG_FLAGS = {"--debug"}
 HELP_FLAGS = {"-help", "--help", "-h", "help"}
 VERBOSE_FLAGS = {"verbose"}
 VERSION_FLAGS = {"--version"}
+SUB_START = "JOB"
+SUB_END = "DONE"