+++ CHANGELOG.md
@@ -5,7 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- Posible headings: Added, Changed, Deprecated, Fixed, Removed, Security -->
 
-## [Unreleased]
+## [Unreleased 0.11.0]
+
+### Added
+
+- Pipeline substitution using `JOB` and `DONE`
 
 ### Changed